+++ CHANGELOG.md
@@ -9,6 +9,8 @@ since 0.4.1, and this project adheres to [Semantic Versioning](https://semver.or
 
 ### Fixed
 
+- Remember user-selected output directory across sessions
+  ([#1123](https://github.com/freedomofpress/dangerzone/issues/1123))
 - Fix default PDF reader selection on Linux ([#814](https://github.com/freedomofpress/dangerzone/issues/814))
 - Update the theme correctly on dark/light mode changes
   ([#1238](https://github.com/freedomofpress/dangerzone/issues/1238))